refactor: respond to PR feedback
renamed module to 'vault' for better extendability
