+++ x/vault/auth/aws_iam_auth.go
@@ -3,6 +3,7 @@ package auth
 import (
 	"context"
 	"fmt"
+	"os"
 
 	"github.com/aws/aws-sdk-go/aws/credentials/stscreds"
 	"github.com/aws/aws-sdk-go/aws/session"
@@ -28,24 +29,26 @@ func NewAWSIamAuth(vaultRole string, roleArn string) *AWSIamAuth {
 	}
 }
 
-func (a *AWSIamAuth) Login(ctx context.Context, client *vaultapi.Client) (*vaultapi.Secret, error) {
-	var awsSession *session.Session
-	var err error
-	if awsSession, err = session.NewSession(); err != nil {
+func (auth *AWSIamAuth) Login(ctx context.Context, client *vaultapi.Client) (*vaultapi.Secret, error) {
+	awsSession, err := session.NewSession()
+	if err != nil {
 		return nil, fmt.Errorf("failed to create AWS session: %w", err)
 	}
-
+	region, ok := os.LookupEnv("AWS_REGION")
+	if !ok {
+		region = defaultStsRegion
+	}
 	loginData, err := awsutil.GenerateLoginData(
-		stscreds.NewCredentials(awsSession, a.roleArn),
+		stscreds.NewCredentials(awsSession, auth.roleArn),
 		"",
-		defaultStsRegion,
+		region,
 		hclog.Default(),
 	)
 	if err != nil {
 		return nil, fmt.Errorf("unable to generate login data for AWS auth endpoint: %w", err)
 	}
 
-	loginData["role"] = a.vaultRole
+	loginData["role"] = auth.vaultRole
 
 	secret, err := client.Logical().WriteWithContext(ctx, loginPath, loginData)
 	if err != nil {
