+++ x/vault/doc.go
@@ -1,7 +1,7 @@
-// Package vaultdecryption adds helper methods for decrypting using vault
+// Package vault adds helper methods for decrypting using vault
 // To Decrypt, you must first create a client and decrypter:
 // settings := client.VaultSettings{
-////				DecrypterRoleArn: <arn here>,
+////				RoleArn: <arn here>,
 ////				VaultAddr:        <vault address here>,
 ////			}
 // client, err :=  client.NewVaultClient(&settings, ctx context.Context)
@@ -9,4 +9,7 @@
 // decrypter := NewVaultDecrypter(client, &settings})
 //
 // decryptedSecret, err := decrypter.Decrypt(keyReferences, encryptedData, ctx)
-package vaultdecryption
+//
+// The default aws region for generating AWS login data is 'us-east-1' which
+// can be changed by using the env var AWS_REGION
+package vault
